\name{spatconnect-package}
\alias{spatconnect}
\alias{spatconnect-package}
\docType{package}
\title{Global and local spatial connectivity for areal data}
\description{
\pkg{spatconnect} implements Betti-0 based global and local indicators of spatial connectivity for scalar areal data represented on a fixed adjacency graph. It includes common-mean global random-relabeling inference, a component-specific activation--merging decomposition, and conditional local random-relabeling inference.
}
\author{
Alvaro Briz-Redon and Marco Ruiz-Valderrama
}
